move compute weights & tests
